pyproject.toml
[tool.poetry]
name = "flutter-server"
version = "0.1.0"
description = "A Flutter development server with hot reload capabilities and web-based code editor"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.1"
flask-cors = "^6.0.1"
requests = "^2.32.3"
anthropic = "^0.54.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
psutil = "^6.1.1"
langfuse = "^3.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

